fastApiProject/requirements.txt
openai~=1.9.0
fastapi~=0.109.0
starlette~=0.36.0
python-dotenv~=1.0.1
passlib~=1.7.4
pydantic~=2.9.2
pandas~=1.3.3
numpy~=1.21.2
pymongo~=4.6.1
python-jose~=3.3.0
uvicorn~=0.26.0
click~=8.1.7
typing-extensions~=4.9.0
h11~=0.14.0
annotated-types~=0.6.0
pydantic-core~=2.23.4
httpx~=0.27.2
sniffio~=1.3.0
idna~=3.6.0
email-validator==2.2.0
google-generativeai
elevenlabs==2.1.0
soundfile ~=0.12.1
motor==3.1.1

mistralai==1.1.0

pyjwt~=2.3.0
bcrypt~=3.2.0
pydantic_settings~=2.2.1
python-multipart~=0.0.5

fpdf

replicate==0.34.2

#dev
pytest-asyncio
pytest
httpx
asgi-lifespan
pillow
httpx
websocket
websockets
websocket-client
